+++ buildpy/CHANGELOG.md
@@ -17,6 +17,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [0.0.1]
 
+- Added `make_relocatable` method for `shared-*` build variants on macOS to make the installed prefix relocatable
+
+- Added cmdline option to add config options `--cfg-opts`
+
+- Changed `cmd` from `subprocess.call` to `subprocess.check_call` which fails on error
+
 - Added Linux support (Python versions 3.11.7 and 3.12.2 tested).
 
 - Added recursive `glob_remove` method for subtractive reduction of the build.